split forms and views (registries)
