+++ lit-any/template-registry/TemplateSelector.js
@@ -1,4 +1,4 @@
-export class TemplateSelector {
+export default class TemplateSelector {
     constructor() {
         this.name = '';
         this._matchers = [];
@@ -16,22 +16,5 @@ export class TemplateSelector {
     // eslint-disable-next-line class-methods-use-this
     shouldMatch() {
         return true;
-    }
-}
-
-export class ViewTemplateSelector extends TemplateSelector {
-    // eslint-disable-next-line class-methods-use-this
-    shouldMatch(criteria) {
-        return typeof criteria.value !== 'undefined' && criteria.value !== null;
-    }
-}
-
-export class FieldTemplateSelector extends TemplateSelector {
-    // eslint-disable-next-line class-methods-use-this
-    shouldMatch(criteria) {
-        return typeof criteria.field !== 'undefined'
-            && criteria.field !== null
-            && typeof criteria.field.property !== 'undefined'
-            && criteria.field.property !== null;
     }
 }
